requirements.txt
Flask==1.1.1
pytest==6.2.3
python-dotenv==0.17.0